--- a/rewrite-gradle/src/test/java/org/openrewrite/gradle/RemoveDependencyTest.java
+++ b/rewrite-gradle/src/test/java/org/openrewrite/gradle/RemoveDependencyTest.java
@@ -28,14 +28,14 @@ import java.util.Optional;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.openrewrite.gradle.Assertions.buildGradle;
 import static org.openrewrite.gradle.Assertions.settingsGradle;
-import static org.openrewrite.gradle.toolingapi.Assertions.withToolingApi;
+import static org.openrewrite.gradle.toolingapi.Assertions.withToolingApiUsingModulesCaching;
 import static org.openrewrite.java.Assertions.mavenProject;
 
 class RemoveDependencyTest implements RewriteTest {
 
     @Override
     public void defaults(RecipeSpec spec) {
-        spec.beforeRecipe(withToolingApi())
+        spec.beforeRecipe(withToolingApiUsingModulesCaching())
           .recipe(new RemoveDependency("org.springframework.boot", "spring-boot*", null));
     }