## What's changed?
Using `withToolingApiUsingModulesCaching` added in https://github.com/openrewrite/rewrite-gradle-tooling-model/pull/33.
Following the advice by @shanman190 at: https://github.com/openrewrite/rewrite/pull/4965#discussion_r1934073362

## What's your motivation?
To speed up CI in this project, specifically the duration of `rewrite-gradle` tests.

EDIT: Adding some non-scientific measurements:

| What | Without this change | With this change |
|------|--------------------|-----------------|
| `./gradlew rewrite-gradle:test` on my laptop | [13m 18s](https://ge.openrewrite.org/s/cvtnsj4koaljw) | [2m 6s](https://ge.openrewrite.org/s/xouyjhlxlyh7g) |
| CI on some other PR touching `rewrite-gradle` only | some PR a few days ago: [9m 43s](https://ge.openrewrite.org/s/27wxickd3r3cw) | this PR: [8m 21s](https://ge.openrewrite.org/s/xyk2hh2c6iwf2) |
| CI on a PR touching `rewrite-core` | [19m 17s](https://ge.openrewrite.org/s/ixitf4ctzeuvs) | [17m 59s](https://ge.openrewrite.org/s/rirclc3feblm4) |

(So it seems like the biggest win is for people living far away from Maven Central servers in terms of network latency)