@knutwannheden 2025-01-31
Looks promising. What difference does that make when running a clean build?