@shanman190 2025-01-31
@greg-at-moderne, the reason that the options exist are because within the CI context and iterating quickly, you want to be able to pick up the latest dynamic version for `latest.integration` for rewrite-maven or the `org.openrewrite.gradle.tooling:plugin` quickly -- which by the way these two options _only_ come into play when pulling from remotely deployed artifacts. When dynamic (latest.release, latest.integration) or changing (1.0-SNAPSHOT) dependencies are in use by default, Gradle will cache these for 24 hours. In OpenRewrite, the snapshots are always used and integrated together rapidly with each other. You can see the same thing being done for all of the top-level OpenRewrite projects via the shared binary build plugin [openrewrite/rewrite-build-gradle-plugin](https://github.com/openrewrite/rewrite-build-gradle-plugin/blob/main/src%2Fmain%2Fjava%2Forg%2Fopenrewrite%2Fgradle%2FRewriteJavaPlugin.java#L46-L49).

So whether you're developing on a core module, the Gradle tooling API, or one of the OSS build plugin, you're getting the latest snapshot versions for everything. It sounds like in your context, you're a little further from one of the Maven Central repositories, so you're experiencing increased latency as a result of these options under all of the above contexts. It may be that you're using the Intellij runner for most activities, which given it bypasses Gradle often, would result in you not feeling the impact of the core module variation conveniently. But when here in rewrite-gradle, it's directly in front of you resulting in being felt early and often.

@jkschneider & @sambsnyd, maybe we've reached a point where a value of `0.seconds` is now becoming a pain point and should be revaluated within the additional data?