@greg-at-moderne 2025-01-31
>  from different contexts and physical locations.

IMHO, our CI running in Github Actions should be the top criterion. I've included my laptop just for the context.