@shanman190 2025-01-31
Gotcha! Good context to know. I'll do some research into optimizing the build based upon the GHA build scans.