@shanman190 2025-02-07
@sambsnyd, I think that 5-10 minutes would be a reasonable value for the default.

One thing I've been pondering is if we flip the script that Greg has started and make it so that the defaults are the 5-10 minutes, then add a `withDevelopmentToolingApi` method which uses an initscript with values of 0. In a local setting though, we're probably just publishing to mavenLocal in which case we get the updated immediately regardless. So having the alternative mechanism may not be worth the test API surface area at all.

But changing the defaults has huge, positive impact across anyone using the tooling API today.