@timtebeek 2025-02-11
Went with the above approach of a five minute cache by default, and pushed a new release that should pick this up:
- https://github.com/openrewrite/rewrite-gradle-tooling-model/releases/tag/v2.13.0

Thanks all!